Moved MessageBox implementation to `.cpp` file
